backend/requirements.txt
aiohttp==3.9.1
aiosignal==1.4.0
aiosqlite==0.19.0
apscheduler==3.10.4
asyncpg==0.29.0
alembic==1.13.1
amqp==5.3.1
annotated-types==0.7.0
anthropic==0.75.0
anyio==4.12.0
attrs==25.4.0
bcrypt==5.0.0
beautifulsoup4==4.12.3
billiard==4.2.3
celery==5.3.6
certifi==2025.11.12
cffi==2.0.0
click==8.3.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
cryptography==46.0.3
distro==1.9.0
docstring_parser==0.17.0
ecdsa==0.19.1
email-validator==2.1.0
fastapi==0.109.0
frozenlist==1.8.0
graphql-core==3.2.7
greenlet==3.0.3
h11==0.16.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.26.0
idna==3.11
jiter==0.12.0
kombu==5.6.1
lia-web==0.2.3
lxml==5.1.0
Mako==1.3.10
MarkupSafe==3.0.3
multidict==6.7.0
packaging==25.0
passlib==1.7.4
playwright==1.41.0
psycopg2-binary==2.9.9
prompt_toolkit==3.0.52
propcache==0.4.1
pyasn1==0.6.1
pycparser==2.23
pydantic==2.5.3
pydantic-settings==2.1.0
pydantic_core==2.14.6
pyee==11.0.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
python-jose==3.3.0
python-multipart==0.0.20
PyYAML==6.0.3
redis==5.0.1
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
soupsieve==2.8
SQLAlchemy==2.0.25
starlette==0.35.1
strawberry-graphql==0.287.2
typing_extensions==4.15.0
tzdata==2025.2
uvicorn==0.27.0
uvloop==0.22.1
vine==5.1.0
watchfiles==1.1.1
wcwidth==0.2.14
websockets==15.0.1
yarl==1.22.0
